+++ src/main/java/firok/tiths/block/BlockBait.java
@@ -1,8 +1,9 @@
 package firok.tiths.block;
 
+import com.zeitheron.hammercore.api.lighting.ColoredLight;
+import com.zeitheron.hammercore.api.lighting.impl.IGlowingBlock;
 import firok.tiths.util.EntityFinders;
 import firok.tiths.util.InnerActions;
-import firok.tiths.util.Predicates;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.resources.I18n;
@@ -12,23 +13,25 @@ import net.minecraft.entity.EntityCreature;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.world.World;
+import net.minecraftforge.fml.common.Optional;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
-import slimeknights.tconstruct.gadgets.TinkerGadgets;
-import slimeknights.tconstruct.gadgets.item.ItemThrowball;
 import slimeknights.tconstruct.shared.block.BlockGlow;
 
-import javax.annotation.Nonnull;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 import java.util.function.Predicate;
 
 @SuppressWarnings("all")
-public class BlockBait extends BlockGlow
+@Optional.Interface(
+		iface = "com.zeitheron.hammercore.api.lighting.impl.IGlowingBlock",
+		modid = "hammercore"
+)
+public class BlockBait extends BlockGlow implements IGlowingBlock
 {
 	Predicate[] canAttracts;
 	@SafeVarargs
@@ -101,5 +104,24 @@ public class BlockBait extends BlockGlow
 	{
 		list.add( I18n.format("tooltip.tiths.block_bait") );
 		InnerActions.addInformation(this,list,flagIn);
+	}
+
+	@Optional.Method(
+			modid = "hammercore"
+	)
+	public ColoredLight produceColoredLight(World world, BlockPos blockPos, IBlockState state, float v) {
+		EnumFacing facing = state.getValue(FACING);
+		int index=facing.getIndex();
+		float r,g,b;
+		switch (index)
+		{
+			case 0: r=0;g=0.5f;b=0.5f; break;
+			case 1: r=1;g=1;b=1; break;
+			case 2: r=0.25f;g=0.7f;b=0.8f;break;
+			case 3: r=0.8f;g=0;b=0; break;
+			default: r=1;g=1;b=0; break;
+		}
+
+		return ColoredLight.builder().pos(blockPos).radius(15.0F).color(r,g,b).build();
 	}
 }